+++ seed/README.md
@@ -1,0 +1,26 @@
+# Seed RPC example
+
+## Execution
+
+Running the server:
+
+```bash
+sbt seed-server/run
+```
+
+Running the client:
+
+```bash
+sbt seed-client/run
+```
+
+(TODO INCLUDE THE RUN COMMANDS FROM THE BUILD.SBT?)
+
+[comment]: # (Start Copyright)
+# Copyright
+
+Mu is designed and developed by 47 Degrees
+
+Copyright (C) 2017-2018 47 Degrees. <http://47deg.com>
+
+[comment]: # (End Copyright)
