+++ seed/client/src/main/scala/example/seed/client/app/AvroClientApp.scala
