+++ seed/client/src/main/scala/example/seed/client/app/ClientBoot.scala
