+++ seed/client/src/main/scala/example/seed/client/app/ProtoClientApp.scala
