+++ seed/client/src/main/scala/example/seed/client/common/models.scala
