--- seed/protocol/modules/protocol_avro/src/main/resources/People.avdl
@@ -1,25 +1,0 @@
-@namespace("example.seed.server.protocol.avro")
-protocol People {
-
-  record Person {
-    string name;
-    int age;
-  }
-
-  record NotFoundError{
-    string message;
-  }
-
-  record DuplicatedPersonError{
-    string message;
-  }
-
-  record PeopleRequest {
-    string name;
-  }
-
-  record PeopleResponse {
-    union{ Person, NotFoundError, DuplicatedPersonError } result;
-  }
-
-}
