--- seed/protocol/modules/protocol_avro/src/main/resources/PeopleService.avdl
@@ -1,7 +1,0 @@
-@namespace("example.seed.server.protocol.avro")
-protocol PeopleService {
-  import idl "People.avdl";
-
-  example.seed.server.protocol.avro.PeopleResponse getPerson(example.seed.server.protocol.avro.PeopleRequest request);
-
-}
