--- seed/protocol/modules/protocol_proto/src/main/scala/people.scala
@@ -1,25 +1,0 @@
-/*
- * Copyright 2020 47 Degrees, LLC. <http://www.47deg.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package example.seed.server.protocol.proto
-
-object people {
-
-  final case class Person(name: String, age: Int)
-  final case class PeopleRequest(name: String)
-  final case class PeopleResponse(person: Person)
-
-}
