+++ seed/protocol/protocol_proto/src/main/resources/people.proto
@@ -1,0 +1,17 @@
+syntax = "proto2";
+
+package example.seed.protocol.proto;
+
+message Person {
+  required string name = 1;
+  required int32 age = 2;
+}
+
+message PeopleRequest {
+  required string name = 1;
+}
+
+message PeopleResponse {
+   required Person person = 1;
+}
+
