+++ seed/server/src/main/scala/example/seed/server/app/ServerBoot.scala
