+++ seed/server/src/main/scala/example/seed/server/common/models.scala
